requirements/dev.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.6.0
Deprecated==1.2.10
distro==1.5.0
Django==3.1.1
django-admin-tools==0.9.0
django-allauth==0.42.0
django-braces==1.14.0
django-cors-headers==3.5.0
django-oauth-toolkit==1.3.2
django-rest-api-logger==0.4.1
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
django-rest-passwordreset==1.1.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djangorestframework-simplejwt-sso==1.0.1
elasticsearch==7.9.1
httplib2==0.18.1
idna==2.10
install==1.3.3
jwt==1.0.0
oauth2==1.9.0.post1
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.0
python-resize-image==1.1.19
python-social-auth @ git+https://github.com/omab/python-social-auth.git@81c0a542d158772bd3486d31834c10af5d5f08b0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.10
uuid==1.30
wrapt==1.12.1
yandex-checkout==1.6.4
